# BACKEND/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Laravel CRUD",
        "description": "Documentação da API Laravel CRUD",
        "contact": {
            "email": "seu-email@dominio.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sales": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Listar todas as vendas com relacionamentos",
                "operationId": "419ebb078c83262b05f7c639ae8b2c21",
                "responses": {
                    "200": {
                        "description": "Lista de vendas retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Criar uma nova venda",
                "operationId": "51f42c056032d143981f26f6c0888f24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_client",
                                    "id_shop",
                                    "id_seller",
                                    "date_sale",
                                    "value_total",
                                    "payment_method",
                                    "products"
                                ],
                                "properties": {
                                    "id_client": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_shop": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "id_seller": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "date_sale": {
                                        "type": "string",
                                        "example": "2025-07-15"
                                    },
                                    "value_total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 199.9
                                    },
                                    "obs": {
                                        "type": "string",
                                        "example": "Venda com desconto"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "Dinheiro",
                                            "Crédito",
                                            "Débito"
                                        ],
                                        "example": "Crédito"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Venda criada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/sales/{sale}": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Visualizar uma venda específica com detalhes",
                "operationId": "e72b118331036aff4e9f004a7b403138",
                "parameters": [
                    {
                        "name": "sale",
                        "in": "path",
                        "description": "ID da venda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Venda retornada com sucesso"
                    },
                    "404": {
                        "description": "Venda não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sales"
                ],
                "summary": "Atualizar informações de uma venda",
                "operationId": "5388b7e083e529ecc6828b1f99da7bf3",
                "parameters": [
                    {
                        "name": "sale",
                        "in": "path",
                        "description": "ID da venda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date_seller",
                                    "value_total",
                                    "payment_method"
                                ],
                                "properties": {
                                    "date_seller": {
                                        "type": "string",
                                        "example": "2025-07-20"
                                    },
                                    "value_total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 299.9
                                    },
                                    "obs": {
                                        "type": "string",
                                        "example": "Atualizado pela gerência"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "Dinheiro",
                                            "Credito",
                                            "Debito"
                                        ],
                                        "example": "Dinheiro"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Venda atualizada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sales"
                ],
                "summary": "Deletar uma venda",
                "operationId": "8381c5f5204ca5aa86f0e97d75a15644",
                "parameters": [
                    {
                        "name": "sale",
                        "in": "path",
                        "description": "ID da venda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Venda deletada com sucesso"
                    },
                    "404": {
                        "description": "Venda não encontrada"
                    }
                }
            }
        },
...
    },
...
}